Test code for task 4.3 is working

+++ tests/test_cleanup.c
@@ -1,7 +1,8 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <time.h>
-#include <hashmap.h>
+
+#include "hashmap.h"
 #include "timer.h"
 
 /* Items inserted / removed from the hashmap */
@@ -11,8 +12,8 @@ const int NUM_ITEMS = 100000;
 const int NUM_LOOKUPS = 10000;
 
 /* Quick validation of the hashmap */
-void validate(HM* hm) {
-    long num = rand();
+void validate(HM *hm) {
+    int num = rand();
 
     if (lookup_item(hm, num) == 0) {
         fprintf(stderr, "Hashmap found item %d when it was not present\n", num);
